refactor: rename edit community page -> edit community info page
